// src/models/docker-image.ts
export interface DockerImage {
  Command: string;
  CreatedAt: string;
  ID: string;
  Image: string;
  Labels: string;
  LocalVolumes: string;
  Mounts: string;
  Names: string;
  Networks: string;
  Ports: string;
  RunningFor: string;
  Size: string;
  State: "running" | "stopped" | "exited" | "created" | "paused";
  Status: string;
}